+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### NEXT
 
+* Node: Migrate tests to TypeScript (PR #958).
 * Node: Remove compiled JavaScript from repository and compile TypeScript code on NPM `prepare` script on demand when installed via git (PR #954).
 * `Worker`: Add `RTC::Shared` singleton for RTC entities (PR #953).